requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-api-python-client==2.86.0
python-dotenv==1.0.0
